tests: split `AppCoreTests` into smaller chunks like other Cores
* Tests are not compilable yet, needs more refinements
